% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coral.R
\docType{data}
\name{coral}
\alias{coral}
\title{Coral trait counts}
\format{
A data frame with 32585 observations on the following 7 variables.
\itemize{
\item{\code{observation_id}}{ }
\item{\code{specie_id}}{ }
\item{\code{specie_name}}{ }
\item{\code{location_name}}{ }
\item{\code{resource_id}}{ }
\item{\code{standard_id}}{ }
\item{\code{standard_unit}}{ }

}
}
\source{
Data collected by Joshua Madin for the Coral Traits database 
See also 
<https://figshare.com/articles/dataset/Coral_Trait_Database_1_1_1/2067414>
}
\usage{
data(coral)
}
\description{
Global estimates for coral species from the Coral Trait Database
}
