% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_pars.R
\name{def_pars}
\alias{def_pars}
\title{Get default model parameters.}
\usage{
def_pars(nsym = 1)
}
\arguments{
\item{nsym}{The number of distinct symbionts to include in the model.}
}
\value{
A list of the default values for
  all parameters used in the model. Each list element is named as the
  parameter name. Host parameters have length one, and symbiont parameters have length \code{nsym}
}
\description{
\code{def_pars} takes returns the default values for all host and symbiont
parameters used in the model.
}
\examples{
def_pars()
pars <- def_pars()
}
