% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_env.R
\name{init_env}
\alias{init_env}
\title{Initialize abiotic environment.}
\usage{
init_env(time, L, N, X)
}
\arguments{
\item{time}{A vector of time steps at which the model should be evaluated
(units=days) (e.g., seq(0, 365, 0.1))}

\item{L}{A vector with length 3 defining external light (min, max, functional
form (see details)). Units=mol photons m^-2 d^-1.}

\item{N}{A vector with length 3 defining external DIN concentration (min,
max, functional form (see details)). Units=mol N L^-1.}

\item{X}{A vector with length 3 defining external prey concentration (min,
max, functional form (see details)). Units=C-mol X L^-1.}
}
\value{
A list of 3 numeric vectors (each with length=length(time))
  corresponding to light, DIN, and prey environmental forcing functions.
}
\description{
This function returns a list of vectors that define the external irradiance
(L), DIN concentration (N), and prey concentration (X) to be used as
subsequent inputs for a model simulation.
}
\details{
The options for the functional form of each environmental factor are:
\itemize{
  \item 1=linear increase from min to max
  \item 2=linear decrease from max to min
  \item 3=sinusoid with period of one year and a range from min to max
}
}
\examples{
env1 <- init_env(time=seq(1,365,0.1), L=c(20,40,1), N=c(1e-7,1e-7,0), X=c(0,0,0))
}
