% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_n.R
\name{mean_n}
\alias{mean_n}
\title{Computing mean over variables and dealing with NAs}
\usage{
mean_n(vars, not.na = 0)
}
\arguments{
\item{vars}{variables (columns) of a data.frame}

\item{not.na}{the minimum number of valid answers (not NA)}
}
\value{
mean
}
\description{
mean_n computes row means from a data.frame or matrix. The first argument of mean_n() is variables (columns of a data frame or matrix) and the second argument a numeric value indicates the minimum amount of valid values (not NA) per row to calculate row mean (default = 1). Numeric value between 0 and 1 indicates a proportion of valid values per row to calculate the row mean.
}
\examples{
\dontrun{
mean_n(df[, 2:5], 3)
mean_n(df[, 1:5], 0.5)
mean_n(mtcars[, c("mpg","cyl")], 2)
}
}
