% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coral_base_analysis.R
\name{plot_gene_omega_curve}
\alias{plot_gene_omega_curve}
\title{Plot Single Gene Omega-squared Flucuation Curve}
\usage{
plot_gene_omega_curve(seurat_obj, gene)
}
\arguments{
\item{seurat_obj}{A Seurat object processed by the CORAL analysis, including
the advanced fluctuation analysis.}

\item{gene}{A string specifying the name of the heritable gene to visualize.}
}
\value{
A ggplot object.
}
\description{
Plots the normalized Omega-squared curve for a single heritable gene
across different levels of clone clustering, comparing it to a reference line.
}
