% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coral_base_analysis.R
\name{plot_heritable_gene_distribution}
\alias{plot_heritable_gene_distribution}
\title{Plot Heritable Gene Effect Size Distribution}
\usage{
plot_heritable_gene_distribution(seurat_obj)
}
\arguments{
\item{seurat_obj}{A Seurat object processed by `run_coral_ground_truth_analysis`.}
}
\value{
A ggplot object.
}
\description{
Creates a density plot comparing the distribution of observed Omega-squared values
  for all genes against the null distribution generated from permutations.
}
\seealso{
\code{\link{run_coral_ground_truth_analysis}}
}
