% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{plot_run}
\alias{plot_run}
\alias{plot_L}
\alias{plot_DIN}
\alias{plot_gr}
\alias{plot_sh}
\alias{plot_Lq}
\alias{plot_photo}
\alias{plot_pl}
\alias{plot_PSU}
\alias{plot_ROS}
\alias{plot_symSU}
\alias{plot_sl}
\alias{plot_corSU}
\alias{plot_hl}
\alias{plot_bl}
\alias{plot_run_v2}
\title{Plot model simulations}
\usage{
plot_run(run)

plot_L(run)

plot_DIN(run)

plot_gr(run)

plot_sh(run)

plot_Lq(run)

plot_photo(run)

plot_pl(run)

plot_PSU(run)

plot_ROS(run)

plot_symSU(run)

plot_sl(run)

plot_corSU(run)

plot_hl(run)

plot_bl(run)

plot_run_v2(run)
}
\arguments{
\item{run}{The model simulation to plot from. This can be the object returned
from \code{run_coral} or \code{run_coral_ss}.}
}
\description{
The \code{plot_run} function will plot a multipanel figure with all model
fluxes. Other functions can be used to plot single environmental factors or
model fluxes independently.
}
\section{Functions}{
\itemize{
\item \code{plot_L}: Plot external irradiance

\item \code{plot_DIN}: Plot external DIN concentration

\item \code{plot_gr}: Plot specific growth rates of host and symbiont

\item \code{plot_sh}: Plot the symbiont to host biomass ratio

\item \code{plot_Lq}: Plot light quenching (carbon fixation, NPQ, and excess (=ROS-producing))

\item \code{plot_photo}: Plot photosynthesis rate (carbon fixation; jCP)

\item \code{plot_pl}: Plot limitation coefficient for photosynthesis SU

\item \code{plot_PSU}: Plot photosynthesis SU dynamics (proportions of light and CO2 rejected from SU)

\item \code{plot_ROS}: Plot relative ROS production due to excitation energy in excess of carbon fixation and NPQ

\item \code{plot_symSU}: Plot symbiont biomass SU dynamics (proportions of C and N rejected from SU)

\item \code{plot_sl}: Plot symbiont biomass SU limitation coefficient

\item \code{plot_corSU}: Plot host biomass SU dynamics (proportions of C and N rejected from SU)

\item \code{plot_hl}: Plot host biomass SU limitation coefficient

\item \code{plot_bl}: Plot both symbiont and host biomass SU limitation coefficients

\item \code{plot_run_v2}: Plot another version of the multipanel run output
}}

