% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coral_base_analysis.R
\name{plot_state_celltype_confusion}
\alias{plot_state_celltype_confusion}
\title{Plot Confusion Matrix of CORAL States vs. Cell Types}
\usage{
plot_state_celltype_confusion(seurat_obj, celltype_col, ...)
}
\arguments{
\item{seurat_obj}{A Seurat object processed by `run_coral_ground_truth_analysis`.}

\item{celltype_col}{A string specifying the metadata column with the cell type or other annotations to compare against.}

\item{...}{Additional parameters passed to `ComplexHeatmap::pheatmap`.}
}
\value{
A `Heatmap` object from the `ComplexHeatmap` package.
}
\description{
Generates a heatmap showing the overlap (confusion matrix) between the
  computed CORAL states and another cell annotation, such as cell type.
}
\seealso{
\code{\link[ComplexHeatmap]{pheatmap}}
}
