% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_coral.R
\name{run_coral}
\alias{run_coral}
\title{Run simulation of coral-\emph{Symbiodinium} model.}
\usage{
run_coral(time, env, pars)
}
\arguments{
\item{time}{A vector of time steps at which the model should be evaluated
(units=days) (e.g., seq(0, 365, 0.1))}

\item{env}{A list of numeric vectors named L, N, and X, defining the external
light, [DIN], and [prey] for the model simulation. Each of these vectors
should have length=length(time). The object returned from \code{init_env}
can be used here.}

\item{pars}{A list of (named) parameter values to use in the simulation.
Parameter names must match those returned by \code{def_pars()}.}
}
\value{
A list with the following named elements:
  \describe{
    \item{time}{The time vector used in the simulation}
    \item{env}{The environment object used in the simulation}
    \item{pars}{The parameter values used in the simulation}
    \item{H}{A tibble of all host biomass-specific model
    fluxes at each time step}
    \item{S}{A tibble of all symbiont biomass-specific
    model fluxes at each time step}
  }
}
\description{
This function runs a simulation of the coral-Symbiodinium bioenergetic model
using a specified time vector, environmental inputs, and parameters.
}
\examples{
time1 <- seq(0, 365, 0.1)
pars1 <- def_pars()
env1 <- init_env(time=time1, L=c(20,40,1), N=c(1e-7,1e-7,0), X=c(0,0,0))
run1 <- run_coral(time=time1, env=env1, pars=pars1)
}
\seealso{
\code{\link{init_env}}, \code{\link{def_pars}}
}
