% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_coral_ss.R
\name{run_coral_ss}
\alias{run_coral_ss}
\title{Run coral-\emph{Symbiodinium} model to steady state.}
\usage{
run_coral_ss(env, pars = def_pars(), dt = 0.1)
}
\arguments{
\item{env}{A list defining the constant environment in which to run model to steady state, with three named elements:
\describe{
    \item{L}{External light in mol photons m^-2 d^-1 (numeric).}
    \item{N}{External [DIN] in mol N L^-1 (numeric).}
    \item{X}{External [prey] in C-mol X L^-1 (numeric).}
  }}

\item{pars}{A list of (named) parameter values to use in the simulation.
Parameter names must match those returned by \code{def_pars()}.}

\item{dt}{The time step (in days) to use in the numerical analysis. Values > 0.1 are not recommended.}
}
\value{
A list with the following named elements:
  \describe{
    \item{env}{A list of the environment (L, N, X) used in the simulation}
    \item{pars}{A list of the parameter values used in the simulation}
    \item{H}{A tibble of all host biomass-specific model
    fluxes at each time step}
    \item{S}{A tibble of all symbiont biomass-specific
    model fluxes at each time step}
  }
}
\description{
This function runs the coral-Symbiodinium bioenergetic model in a constant
environment until it reaches a steady state.
}
\examples{
ssrun <- run_coral_ss(env=list(L=20, N=1e-7, X=0), pars=def_pars(), dt=0.1)
}
\seealso{
\code{\link{init_env}}, \code{\link{def_pars}}
}
