% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym_PI.R
\name{sym_PI}
\alias{sym_PI}
\title{Simulate and plot a \emph{Symbiodinium} PI curve.}
\usage{
sym_PI(pars)
}
\arguments{
\item{pars}{A list of model parameters to use in the simulation, i.e. the
object returned by \code{def_pars}.}
}
\value{
A data frame with simulation results at each time step.
}
\description{
Given a set of parameters defining \emph{Symbiodinium} photobiology (e.g.,
yCL, jCPm, kROS, kNPQ, k), \code{sym_PI} calculates and plots the rate of
photosynthesis (jCP), NPQ (jNPQ), and ROS production (cROS) as a function of
irradiance. This simulation is meant to show the photosynthetic performance
of a given symbiont independent from a host (i.e., in a state that is not
influenced by host carbon-concentrating mechanisms), so the "jCO2" parameter
is set at a very high value.
}
\examples{
PI1 <- sym_PI(def_pars())

}
