% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth.R
\name{synth}
\alias{synth}
\title{Parallel complementary synthesizing unit.}
\usage{
synth(x, y, m)
}
\arguments{
\item{x, y}{The two substrate input fluxes to the synthesizing unit as
length-one numerics.}

\item{m}{The maximum production flux of the synthesizing unit.}
}
\value{
The production flux as a length-one numeric.
}
\description{
Given two substrate input fluxes, \code{synth} calculates the production flux
of a parallel complementary synthesizing unit. See page 105, Fig. 3.7, of:
Kooijman, S. A. L. M. (2010). Dynamic Energy Budget Theory for Metabolic
Organisation. Cambridge University Press. 3rd edition.
}
\examples{
synth(1, 1, 10)
}
