% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varlist.R
\name{varlist}
\alias{varlist}
\title{List variables}
\usage{
varlist(d)
}
\arguments{
\item{d}{a data frame}
}
\value{
a data frame (rows are variables and columns are variable attributes)
}
\description{
List variables
}
\examples{
\dontrun{
varlist(df)  # df is a data frame
}
}
