% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varlist_shiny.R
\name{varlist.shiny}
\alias{varlist.shiny}
\title{List variables in the browser with shiny}
\usage{
varlist.shiny(d)
}
\arguments{
\item{d}{a data frame}
}
\value{
a data frame
}
\description{
List variables in the browser with shiny
}
\examples{
\dontrun{
varlist.shiny(df)  # df is a data frame
}
}
