% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varlist_view.R
\name{varlist.view}
\alias{varlist.view}
\title{List variables in the viewer}
\usage{
varlist.view(d)
}
\arguments{
\item{d}{a data.frame}
}
\value{
a data.frame
}
\description{
List variables in the viewer
}
\examples{
\dontrun{
varlist.view(df)
}
}
