% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coral_base_analysis.R
\name{visualize_clone_distance_heatmap}
\alias{visualize_clone_distance_heatmap}
\title{Visualize the Clone Energy Distance Heatmap}
\usage{
visualize_clone_distance_heatmap(seurat_obj, ...)
}
\arguments{
\item{seurat_obj}{A Seurat object processed by `run_coral_ground_truth_analysis`.}

\item{...}{Additional parameters passed to `ComplexHeatmap::Heatmap` for customization.}
}
\value{
A `Heatmap` object from the `ComplexHeatmap` package.
}
\description{
Creates a heatmap of the clone-clone energy distance matrix, visualizing
  the relationships between clones and the defined CORAL states.
}
\seealso{
\code{\link{run_coral_ground_truth_analysis}}, \code{\link[ComplexHeatmap]{Heatmap}}
}
