% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coral_base_analysis.R
\name{visualize_clone_mds}
\alias{visualize_clone_mds}
\title{Visualize the MDS Embedding of Clones}
\usage{
visualize_clone_mds(
  seurat_obj,
  color_by = "coral_state",
  mds_dims = c(1, 2),
  pt_size = 2
)
}
\arguments{
\item{seurat_obj}{A Seurat object processed by `run_coral_ground_truth_analysis`.}

\item{color_by}{A string specifying how to color the points. Can be `"coral_state"` (default)
or any valid column name from the Seurat object's metadata.}

\item{mds_dims}{A numeric vector of length 2 specifying which MDS dimensions to plot. Defaults to `c(1, 2)`.}

\item{pt_size}{The size of the points in the plot. Defaults to `2`.}
}
\value{
A ggplot object.
}
\description{
Generates a 2D scatter plot of clones based on the Multi-Dimensional Scaling (MDS)
  of the energy distance matrix.
}
\seealso{
\code{\link{run_coral_ground_truth_analysis}}
}
