% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coral_base_analysis.R
\name{visualize_coral_states_umap}
\alias{visualize_coral_states_umap}
\title{Visualize CORAL States on UMAP}
\usage{
visualize_coral_states_umap(seurat_obj, ...)
}
\arguments{
\item{seurat_obj}{A Seurat object processed by `run_coral_ground_truth_analysis`.}

\item{...}{Additional parameters passed to `Seurat::DimPlot`.}
}
\value{
A ggplot object.
}
\description{
Overlays the calculated CORAL states on a UMAP projection of the cells.
  Note: Requires UMAP to be computed on the Seurat object first.
}
\seealso{
\code{\link[Seurat]{DimPlot}}, \code{\link{run_coral_ground_truth_analysis}}
}
