% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coral_base_analysis.R
\name{visualize_gene_mds}
\alias{visualize_gene_mds}
\title{Visualize Gene Expression on Clone MDS Plot}
\usage{
visualize_gene_mds(
  seurat_obj,
  gene,
  log_transform = TRUE,
  mds_dims = c(1, 2),
  pt_size = 2
)
}
\arguments{
\item{seurat_obj}{A Seurat object processed by the CORAL analysis.}

\item{gene}{A string specifying the name of the gene to visualize.}

\item{log_transform}{Logical, whether to log2(x+1) transform the expression
values for visualization. Defaults to TRUE.}

\item{mds_dims}{A numeric vector of length 2 indicating which MDS dimensions to plot.
Defaults to `c(1, 2)`.}

\item{pt_size}{The size of the points in the plot.}
}
\value{
A ggplot object.
}
\description{
Overlays the pseudobulk expression of a specific gene onto the
MDS embedding of clones. The required pseudobulk matrix is now calculated internally.
}
