% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_pars.R
\name{def_pars}
\alias{def_pars}
\title{Get default model parameters.}
\usage{
def_pars()
}
\value{
A list of 26 length-one numerics containing the default values for
  all parameters used in the model. Each list element is named as the
  parameter name.
}
\description{
\code{def_pars} takes no arguments and returns the default values for all
parameters used in the model.
}
\examples{
def_pars()
pars <- def_pars()
}
