% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_n.R
\name{mean_n}
\alias{mean_n}
\title{Computing mean over variables and dealing with NAs}
\usage{
mean_n(vars, not.na = 0)
}
\arguments{
\item{vars}{variables (columns) of a data.frame}

\item{not.na}{the minimum number of valid answers (not NA)}
}
\value{
mean
}
\description{
Computing mean over variables and dealing with NAs
}
\examples{
\dontrun{
mean_n(df[, 2:5], 3)
}
}
