% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coral_base_analysis.R
\name{plot_gene_dashboard}
\alias{plot_gene_dashboard}
\title{Create a Comprehensive Gene Analysis Dashboard}
\usage{
plot_gene_dashboard(seurat_obj, gene, num_ridge_idents = 12)
}
\arguments{
\item{seurat_obj}{A Seurat object processed by the CORAL analysis.}

\item{gene}{A string specifying the name of the gene to analyze.}

\item{num_ridge_idents}{An integer for the number of top clones to show in the
ridge plot. Defaults to 12.}
}
\value{
A plot object created by `ggpubr::ggarrange`.
}
\description{
Combines multiple plots for a single gene into one comprehensive view.
It now includes an AUC value annotation on the fluctuation curve plot.
The required pseudobulk matrix is calculated internally.
}
